+++ game/behaviours.go
@@ -1,0 +1,9 @@
+package game
+
+func shouldAttack() bool {
+	return false
+}
+
+func shouldEatFood() bool {
+	return true
+}
